msm: camera: sensor: Add CCI support for 32 bits register Address
Add support for 32 bits wide sensor ID for CCI hw during sensor
match ID operation.

CRs-Fixed: 3210196

+++ drivers/cam_sensor_module/cam_sensor_utils/cam_sensor_cmn_header.h
@@ -262,10 +262,10 @@ struct cam_sensor_power_ctrl_t {
 };
 
 struct cam_camera_slave_info {
-	uint16_t sensor_slave_addr;
-	uint16_t sensor_id_reg_addr;
-	uint16_t sensor_id;
-	uint16_t sensor_id_mask;
+	uint32_t sensor_slave_addr;
+	uint32_t sensor_id_reg_addr;
+	uint32_t sensor_id;
+	uint32_t sensor_id_mask;
 };
 
 struct msm_sensor_init_params {
